requirements.txt
flask-script
flask-migrate
flask-rest-api
marshmallow == 3.0.0b19
marshmallow-sqlalchemy
marshmallow-enum
psycopg2
